requirements.txt
aiohttp==3.9.0
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.0
async-timeout==4.0.3
attrs==23.1.0
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pangres==4.2.1
parsel==1.8.1
protobuf==4.21.12
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
w3lib==2.1.2
yarl==1.9.3
